move get total votes and get class names for buttns functions into a file
